Renamed most types introduced in this work and filled out GoDoc.
